# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        description: not string hashed password
        type: string
      role:
        description: Role ? Admin OR manager - one role for now - admin
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Your ultimate destination for all things fresh, flavorful, and fabulous
    – where convenience meets quality in our one-stop grocery wonderland!
  title: "One Stop Grocery\U0001F6D2"
  version: "1.0"
paths:
...
  /userLogin:
    post:
      consumes:
      - application/json
      description: Authenticates a user with the provided email and password, returning
        a JWT token upon success
      operationId: login-user
      parameters:
      - description: Login credentials (email and password)
        in: body
        name: loginCredentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Failed to generate token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Log in a user
  /users:
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided information
      operationId: create-new-user
      parameters:
      - description: User object to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User Created Successfully. userID
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request format" or "Missing fields in the request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error" or "Failed to create user in Firestore
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new user
schemes:
- http
swagger: "2.0"